fix signle video menu
