+++ src/App.css
@@ -22,6 +22,6 @@
     width: 100%;
     height: auto;
     top: auto;
-    z-index: 3;
+    z-index: 4;
   }
 }
