+++ src/css/utility.css
@@ -97,3 +97,7 @@
   display: flex;
   justify-content: space-between;
 }
+
+.pos-rel {
+  position: relative;
+}
