+++ src/pages/single-video-page/Single-video-page.jsx
@@ -1,7 +1,7 @@
 import { useParams } from "react-router-dom";
 import "./single-video-page.css"
 import { useVideoDataFromServer } from "../../context/video-context"
-import { VideoPlayer, VideoPlayerFooter } from './components';
+import { VideoPlayer, VideoPlayerCentre } from './components';
 import { VideoCard } from "../../components";
 
 export const SingleVideoPage = () => {
@@ -16,7 +16,7 @@ export const SingleVideoPage = () => {
                 videoState.allVideos.length > 0 ? <div>
                     <div className='single-video-player'>
                         <VideoPlayer video={currentVideo} />
-                        <VideoPlayerFooter currentVideo={currentVideo} />
+                        <VideoPlayerCentre currentVideo={currentVideo} />
                     </div>
                     <div>
                     </div>
