+++ src/pages/single-video-page/components/index.js
@@ -1,2 +1,2 @@
 export {VideoPlayer} from "./VideoPlayer"
-export {VideoPlayerFooter} from "./VideoPlayerFooter"
+export {VideoPlayerCentre} from "./VideoPlayerCentre"
