+++ src/pages/single-video-page/single-video-page.css
@@ -17,7 +17,7 @@
   height: 30rem;
 }
 
-.single-video-footer {
+.single-video-centre {
   display: flex;
   justify-content: space-between;
   padding: 1rem;
